Title: 특정 숫자 도달하기, Time: 124ms, Memory: 10MB, Status: Passed - CodeTree

+++ 240723/특정 숫자 도달하기/reaching-specific-number.java
@@ -17,7 +17,7 @@ public class Main {
             sum += a;
         }
         if(!check){
-            System.out.printf("%d %.1f", sum, sum/10);
+            System.out.printf("%d %.1f", sum, (double)sum/10);
         }
 
     }
